Make/model as entities

+++ RentACar.Core/DTOs/CarModelDTO.cs
@@ -8,13 +8,13 @@ namespace RentACar.Core.DTOs
     public class CarModelDTO
     {
         public int Id { get; set; }
-        [Range((int)1, int.MaxValue, ErrorMessage = "CarMakeId must be at least 1 (and point to a valid car make entity)")]
+        [Range(1, int.MaxValue, ErrorMessage = "CarMakeId must be at least 1 (and point to a valid car make entity)")]
         public int CarMakeId { get; set; }
 
 
         [MinLength(1)]
         [MaxLength(100)]
-        string LogoFilename { get; set; } // optional
+        string LogoFilename { get; set; } 
 
 
         [MinLength(1)]
